Former-commit-id: 633c3c0b2a7772eee1107695bacb136289590778

+++ hq-server/src/main/java/org/hyperic/hq/escalation/server/session/EscalationRuntimeImpl.java
@@ -272,7 +272,9 @@ public class EscalationRuntimeImpl implements EscalationRuntime {
 
 			}
 		}
-		escalationStateRepository.deleteByIds(stateIds);
+		if(!stateIds.isEmpty()) {
+		    escalationStateRepository.deleteByIds(stateIds);
+		}
 	}
 
 	private void unscheduleEscalation_(Integer stateId) {
